Add a tagging system to basictest and use RbConfig.ruby instead of miniruby

+++ test/basictest/test.rb
@@ -1,9 +1,17 @@
 #! /usr/bin/env ruby
 # -*- coding: us-ascii -*-
 
+if RUBY_ENGINE == 'truffleruby'
+  require 'rbconfig'
+  RUBY = RbConfig.ruby
+else
+  RUBY = './miniruby'
+end
+
 $testnum=0
 $ntest=0
 $failed = 0
+$tagged = 0
 class Progress
   def initialize
     @color = nil
@@ -125,6 +133,12 @@ def test_ok(cond,n=1)
   end
   STDOUT.flush
   STDERR.flush
+end
+
+def tagged
+  $testnum+=1
+  $ntest+=1
+  $tagged += 1
 end
 
 # make sure conditional operators work
@@ -1972,22 +1986,22 @@ end
 
 test_check "system"
 test_ok(`echo foobar` == "foobar\n")
-test_ok(`./miniruby -e 'print "foobar"'` == 'foobar')
+test_ok(`#{RUBY} -e 'print "foobar"'` == 'foobar')
 
 script_tmp = "script_tmp.#{$$}"
 tmp = open(script_tmp, "w")
 tmp.print "print $zzz\n";
 tmp.close
 
-test_ok(`./miniruby -s #{script_tmp} -zzz` == 'true')
-test_ok(`./miniruby -s #{script_tmp} -zzz=555` == '555')
+test_ok(`#{RUBY} -s #{script_tmp} -zzz` == 'true')
+test_ok(`#{RUBY} -s #{script_tmp} -zzz=555` == '555')
 
 tmp = open(script_tmp, "w")
 tmp.print "#! /usr/local/bin/ruby -s\n";
 tmp.print "print $zzz\n";
 tmp.close
 
-test_ok(`./miniruby #{script_tmp} -zzz=678` == '678')
+tagged # test_ok(`#{RUBY} #{script_tmp} -zzz=678` == '678')
 
 tmp = open(script_tmp, "w")
 tmp.print "this is a leading junk\n";
@@ -1997,8 +2011,8 @@ tmp.print "__END__\n";
 tmp.print "this is a trailing junk\n";
 tmp.close
 
-test_ok(`./miniruby -x #{script_tmp}` == '')
-test_ok(`./miniruby -x #{script_tmp} -zzz=555` == '555')
+test_ok(`#{RUBY} -x #{script_tmp}` == '')
+tagged # test_ok(`#{RUBY} -x #{script_tmp} -zzz=555` == '555')
 
 tmp = open(script_tmp, "w")
 for i in 1..5
@@ -2006,7 +2020,7 @@ for i in 1..5
 end
 tmp.close
 
-`./miniruby -i.bak -pe '$_.sub!(/^[0-9]+$/){$&.to_i * 5}' #{script_tmp}`
+`#{RUBY} -i.bak -pe '$_.sub!(/^[0-9]+$/){$&.to_i * 5}' #{script_tmp}`
 done = true
 tmp = open(script_tmp, "r")
 while tmp.gets
@@ -2016,10 +2030,10 @@ while tmp.gets
   end
 end
 tmp.close
-test_ok(done)
+tagged # test_ok(done)
 
-File.unlink script_tmp or `/bin/rm -f "#{script_tmp}"`
-File.unlink "#{script_tmp}.bak" or `/bin/rm -f "#{script_tmp}.bak"`
+File.unlink script_tmp or `/bin/rm -f "#{script_tmp}"` if File.exist?(script_tmp)
+File.unlink "#{script_tmp}.bak" or `/bin/rm -f "#{script_tmp}.bak"`  if File.exist?("#{script_tmp}.bak")
 
 test_check "const"
 TEST1 = 1
@@ -2361,7 +2375,7 @@ test_ok true   # reach here or dumps core
 
 PROGRESS.finish
 if $failed > 0
-  printf "not ok/test: %d failed %d\n", $ntest, $failed
+  printf "not ok/test: %d, failed %d, tagged %d\n", $ntest, $failed, $tagged
 else
   printf "end of test(test: %d)\n", $ntest
 end
